refactor: validation, error handlers
