+++ src/errorHandlers/handle-file-error.js
